Change the setting name to Home and add order button to Home page, and remove it from navbar
