+++ lib/Screens/Home.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:firebase_auth/firebase_auth.dart';
- import 'package:pl2/Screens/Signin.dart';
+import 'package:pl2/Screens/Signin.dart';
 import 'package:pl2/Screens/MyProfile.dart';
 import 'package:pl2/Screens/History.dart';
 import 'package:pl2/Payments/PaymentMethod.dart';
+import 'package:pl2/Screens/Order.dart';
 
+//Home Page 
 class SettingPage extends StatefulWidget{
   @override
   _SettingPage createState() => _SettingPage();
@@ -24,6 +26,7 @@ class _SettingPage extends State<SettingPage>{
           height: MediaQuery.of(context).size.height,
           padding: EdgeInsets.symmetric(horizontal: 30, vertical:30),
             child:Column(
+              //Order Button
               children:<Widget> [
                 Padding(
                   padding:  EdgeInsets.only(top: 30),
@@ -31,7 +34,7 @@ class _SettingPage extends State<SettingPage>{
                     minWidth:double.infinity,
                     height: 60,
                     onPressed: (){
-                       Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (context) => MyProfilePage()));
+                       Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (context) => BodyMainPage()));
                     },
                     color:Colors.white,
                     shape:RoundedRectangleBorder(
@@ -40,10 +43,34 @@ class _SettingPage extends State<SettingPage>{
                           color: Colors.black
                       ),
                     ),
-                      child:Text("My Profile")
+                      child:Text("Order")
                   ),
                 ),
 
+                //MyProfile Button                   
+                    Column(
+                      children:<Widget>[
+                        FlatButton(
+                          minWidth: double.infinity,
+                          height:60,
+                          onPressed: (){
+                              Navigator.of(context).pushReplacement(
+                                MaterialPageRoute(
+                                  builder: (context) =>MyProfilePage()
+                              )
+                            );
+                          },
+                          color:Colors.white,
+                          shape:RoundedRectangleBorder(
+                            borderRadius: (BorderRadius.circular(15)),
+                            side:BorderSide(
+                              color:Colors.black
+                            ),
+                          ),
+                            child:Text("MyProfile")
+                        ),
+                      ],
+                    ),
                     //Payment Method Button
                     Column(
                       children:<Widget>[
@@ -65,6 +92,7 @@ class _SettingPage extends State<SettingPage>{
                       ],
                     ),
 
+                    //History
                     Column(
                       children:<Widget>[
                         FlatButton(
@@ -88,6 +116,8 @@ class _SettingPage extends State<SettingPage>{
                         ),
                       ],
                     ),
+
+                    
                 Column(
                   children: <Widget>[
                     FlatButton(
